feat: export basic limiter config
This is to avoid extra complexity in consumers that expose this config
to users through other means or want to build the config
programmatically, but are currently required to redefine these structs
and keep them in sync. The coupling is already there, we might as well
export these structs and reduce maintenance costs.